#0: Moved input preprocess to llama common
